+++ eslint.config.mjs
@@ -1,0 +1,42 @@
+import { dirname } from "path";
+import { fileURLToPath } from "url";
+import { FlatCompat } from "@eslint/eslintrc";
+import unusedImports from "eslint-plugin-unused-imports";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const compat = new FlatCompat({
+  baseDirectory: __dirname,
+});
+
+const nextConfig = compat.extends("next/core-web-vitals");
+
+for (const config of nextConfig) {
+  if (config.plugins && config.plugins.react && config.plugins.react.configs) {
+    delete config.plugins.react.configs;
+  }
+}
+
+// 3. 최종 무장 완료
+export default [
+  ...nextConfig,
+  {
+    plugins: {
+      "unused-imports": unusedImports,
+    },
+    rules: {
+      "no-unused-vars": "off",
+      "unused-imports/no-unused-imports": "error",
+      "unused-imports/no-unused-vars": [
+        "warn",
+        {
+          vars: "all",
+          varsIgnorePattern: "^_",
+          args: "after-used",
+          argsIgnorePattern: "^_",
+        },
+      ],
+    },
+  },
+];