+++ src/features/post/ui/detail/element/MoreContentSection.tsx
@@ -10,9 +10,9 @@ const MoreContentSection = ({ morePosts }: Props) => {
   return (
     <section className="w-full py-10  md:px-8 lg:px-16 bg-background">
       <div className="text-center mb-8">
-        <h3 className="text-xl sm:text-2xl font-semibold text-foreground">
+        <p className="text-xl sm:text-2xl font-semibold text-foreground">
           다른 게시글도 살펴보세요!
-        </h3>
+        </p>
         <p className="text-sm sm:text-base text-muted-foreground mt-1">
           혹시 도움이 될지도 모르잖아요..?
         </p>